Realizado hasta el modulo de Cotizacion

+++ app/crud/cliente.py
@@ -1,5 +1,8 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException
 from app.models.cliente import Cliente
+from app.models.venta import Venta
+from app.models.cotizacion import Cotizacion
 from app.schemas.cliente import ClienteCreate
 from app.utils import capitalizar
 
@@ -35,6 +38,10 @@ def update_cliente(db: Session, cliente_id: int, cliente: ClienteCreate):
 def delete_cliente(db: Session, cliente_id: int):
     db_cliente = get_cliente(db, cliente_id)
     if db_cliente:
+        en_ventas = db.query(Venta).filter(Venta.cliente_id == cliente_id).first()
+        en_cotizaciones = db.query(Cotizacion).filter(Cotizacion.cliente_id == cliente_id).first()
+        if en_ventas or en_cotizaciones:
+            raise HTTPException(status_code=400, detail="No se puede eliminar porque tiene ventas o cotizaciones asociadas")
         db.delete(db_cliente)
         db.commit()
     return db_cliente
